+++ bgpd/bgp_aspath.h
@@ -58,7 +58,7 @@ struct aspath {
 	/* String expression of AS path.  This string is used by vty output
 	   and AS path regular expression match.  */
 	char *str;
-	unsigned short str_len;
+	size_t str_len;
 	uint32_t count;
 
 	/* AS notation used by string expression of AS path */
